+++ convex-core/src/main/java/convex/core/data/type/IndexType.java
@@ -2,7 +2,9 @@ package convex.core.data.type;
 
 import convex.core.data.ACell;
 import convex.core.data.AIndex;
+import convex.core.data.AString;
 import convex.core.data.Index;
+import convex.core.data.Strings;
 
 /**
  * Type that represents any CVM map
@@ -11,6 +13,8 @@ import convex.core.data.Index;
 public class IndexType extends AStandardType<AIndex> {
 
 	public static final IndexType INSTANCE = new IndexType();
+	
+	private static final AString TAG=Strings.create("#Index");
 	
 	private IndexType() {
 		super(AIndex.class);
@@ -35,6 +39,11 @@ public class IndexType extends AStandardType<AIndex> {
 	public AIndex implicitCast(ACell a) {
 		if (a instanceof Index) return (Index)a;
 		return null;
+	}
+	
+	@Override
+	public AString getTag() {
+		return TAG;
 	}
 
 }