+++ convex-core/src/test/java/convex/core/data/IndexTest.java
@@ -233,6 +233,11 @@ public class IndexTest {
 		assertNotEquals(Index.create(bb, bl), Index.create(ba,bl)); // different entry key types
 		assertEquals(Index.create(bb, bl), Index.create(bl,bl)); // same entry key types
 	}
+	
+	@Test 
+	public void testPrint() {
+		assertEquals("#Index {}",Index.EMPTY.toString());
+	}
 
 	@Test
 	public void testSingleEntry() throws InvalidDataException {
@@ -441,6 +446,9 @@ public class IndexTest {
 			MapEntry<K, V> e2 = m.entryAt(n - 1);
 			assertTrue(e1.getKey().compareTo(e2.getKey().toBlob()) < 0);
 		}
+		
+		// TODO: should round trip when all child types do
+		// assertEquals(m,Reader.read(RT.print(m).toString()));
 		
 		assertEquals(Types.INDEX,m.getType());